Convert cluster aggregator into an equivalence class that must be supplied by cost model, instead of creating it implicitly.
This change makes it the responsibility of cost models to implement a
default EC that contains all machines (or racks, or whatever). This
increases flexibility, as we no longer mandate the existence of a
cluster aggregator node. This commit only makes the necessary changes
to FlowGraph and other system components; a subsequent commit will
add support to the cost model.s

Affected modules: engine misc scheduling/flow

Change-Id: I0e7f80c0227caa587cb766b4ddfc8b832a845087