Update patch set 1
Patch Set 1:

> Unsure about semantics for GetEquivClasstToEquivClassesArcs. Seems
 > to require either a pair <NULL,NULL> or a pair of two vectors. I'd
 > favour considering each element individually, so you can have
 > <NULL, vector> and <vector, NULL> if you want. If you really do
 > want to exclude this case, then should add CHECK_NOTNULL, as
 > currently the program will just segfault.

The conditional on line 135 is correct, I believe, since if either of the two vectors is NULL, there is nothing to do. If I have no sources for my arcs, I won't have arcs, and if I don't have destinations, I won't have any either. That said, this should never happen and would itself be a violation of the API (you cannot have an arc with only a source or only a destination). So maybe we should fail more aggressively, rather than just silently dropping the call on the floor?

I don't see where an additional CHECK_NOTNULL would be required, though If either of the pointers is NULL or both are, we exit the call, so no pointer is dereferenced.

Patch-set: 1
Label: Code-Review=0