Implement Whare-M policy in WhareMap cost model.
This change adds all necessary functionality to support the Whare-M cost
model: when a task finishes, the psPI value (picoseconds per
instruction, an inverse version of Whare-Map's instructions per second,
IPS, since the flow network requires greater cost to correspond to worse
placements) is stored under a mapping of the form

 <task equiv class, machine equiv class> -> [psPI val 0, ..., psPI val n]

This requires a number of API changes in the cost model and in other
parts of the system. Specifically, the TD now also stores copies of
the scheduled resource and the task's final report (for easier reference
from the cost model), which are stored by the EventDrivenScheduler base
once a task has been placed and once it completes.

There is more work to be done: the pspi_map_ contents in the WhareMap
cost model are not currently used; this will follow in a separate change.

Affected modules: base engine scheduling scheduling/flow

Change-Id: I84bb6829f79443495e267e038c9d9b5d8fa9365e