Workaround for cs2 failing with sparse node IDs in flow network.
This commit brings in a somewhat hacky workaround for the problem
described in #25: when using the cs2 solver, the maximum node ID in the
output can be no greater than the number of nodes in the flow graph. The
workaround fix is to return the maximum ID in use instead of the true
number of nodes, which leads to cs2 implicitly assuming disconnected
nodes for the unused IDs.

This commit does not currently affect correctness of any code, but care
is required, since FlowGraph::NumNodes() does not necessarily return the
true number of nodes any more.

Affected modules: scheduling/flow
Bug: #25

Change-Id: I5de0b35cc84f1d4e53e050ac96e25924d031c7ae