Update patch set 1
Patch Set 1:

Ionel, Adam: please check that the repeated call of AddOrUpdateJobNodes() is fine and generates the correct deltas. I have verified this by testing with Flowlessly and cs2, and I believe that this is correct, as we use the same code path for spawning tasks. However, I'd be glad if you could approve that the appropriate arc change deltas are generated.

Patch-set: 1
Label: Code-Review=0