Update patch set 1
Patch Set 1: Code-Review+1

Looks good to me. 

Only point I'd note is that we might at a future date want to update different jobs with different time periods. For example, for a job that's just started, it's probably important to update the preemption cost frequently. But when a job has been running for 3600s, it's unlikely to make much difference whether we update it 1s or 30s. This optimisation would be premature at this point, though.

Patch-set: 1
Label: Code-Review=+1