Gracefully recover from parent coordinator failure by reconnecting.
This change makes StreamSocketChannels partially tolerant to
failures in the underlying transport. Specifically, if an
underlying TCP connection drops because the remote end went away
(e.g., due to a parent coordinator crash), the channel attempts to
repair the connection. If the repair fails, we retry on the next
attempt to use the channel for I/O (and return an error indication).

In addition, and building upon this facility, this change makes
subordinate coordinators continually attempt to re-establish their
connection to their parent. If the connection recovers, we send a
new registration message. This is fine even in the case of a
transient outage without state loss at the parent coordinator, as
the registration is an idempotent action.

Affected modules: engine platforms/unix

Change-Id: I39ebbd3a188f56d8c49ad69acba0660a4b6893de