Use correct worst-case costs in Whare-M.
The Whare-M cost model requires that the cost to the cluster aggregator
be set to the worst-case PsPI value over all machine ECs. This change
tracks the average PsPI value for each machine EC, and stores the
worst-case values for each task EC. This implements the equation

 alpha_{c_t} = max_{c_m}(Psi(c_t, c_m)))

from the cost model description.

Affected modules: scheduling/flow

Change-Id: Idbdf6b1b95c00ed4bd6bb2a5b20aa2ab24d01641