Add coordinator node in topology even if no local resources exist.
This change fixes a bug related to the "virtual" coordinator node in
the resource topology not being registered correctly if the
--include_local_resources flag was set to "false". In such a situation,
AddResource would not be invoked on the (nevertheless existing) node
representing the coordinator. After child resources were registered, the
cost models' statistics update functionality would then encounter this
seemingly unknown resource node and cause a failure.

The solution is to call AddResource on the singular node on startup if
no local resource topology exploration is done.

Affected modules: engine

Change-Id: I67c0622d6ef6805af2cd1310de7f166361df9772