Correct insertion of remote tasks' final reports to parent coordinator.
Before this change, the delivery of delegated tasks' final reports was
implemented by adding a second submessage into the message informing the
parent coordinator of the task completion (both TaskStateMessage and
TaskFinalReport were present). However, since the TaskFinalReport part
of the message was only processed *after* the task state change, the cost
model's update functions were invoked before the report had been
processed, and thus no report was linked in the TD yet. This commit adds
an optional field for the report to the TaskStateMessage, allowing it to
be processed alongside that.

Affected modules: engine messages

Change-Id: I90e02b2166a2378eecec5e28355b0b40d771a8f4