Correctly assign arc capacities between EC aggregators.
This change addresses a TODO on setting the capacities between EC
aggregators correctly (they were previously always 1). We now set
the capacity to the max of the source EC's incoming capacity and
the destination EC's outgoing capacity. This works, although it's
not optimal: we could use the min, to give tighter bounds to the
solver, but doing so would require us to dynamically update the
capacities at runtime, which we currently don't.

Such dynamic updates may, however, still be required even with the
current model when more than two layers of ECs are connected.

Affected modules: scheduling/flow

Change-Id: I3c3cb8072d74277e5930d97ca803fefccb6ef9f1