Skip not-yet-existent tasks in AddEquivClassForNode.
This is a small, but important change: when submitting a job with
more than one task, a situation may arise in which not all tasks in
an EC have been added to the flow graph yet, but we already start
connecting up the EC aggregator. It is fine (with current assumptions)
to skip the not-yet-existent tasks in this case; but see the comment
in the code for details.

Affected modules: scheduling/flow

Change-Id: I0d7bc82ef74fd2a77a196fe585fa53cd6410c4ad