Skip calculation of PsPI in Whare-M cost model if no instruction count is available.
If a task does not report its instruction count (either because
performance monitoring is disabled, or because the data were lost),
we would previously hit a division by zero. This change makes the
computation and recording of the PsPI value conditional on the
instruction count being present.

Affected modules: scheduling/flow

Change-Id: If918a02aa6eb52fcf4dc7769713f77e1b651670b