Force naive load-balancing across machines in Whare-Map cost model.
This change introduces a simple load balancing notion to the Whare-Map
cost model: the number of tasks on each machine should be balanced as
long as long as other machines with fewer utilized cores are available.
We achieve this by setting the cost on the final arc in the resource
topology to the core ID. However, to avoid off behaviour, we also have
to scale all other costs up by the maximum core ID in the cluster now.

Affected modules: scheduling/flow

Change-Id: I5201fdb2ce328704dd97bd0418e8d7ec27772ddc