Update patch set 1
Patch Set 1:

> LGTM. However, this diff highlights an interesting limitation: we
 > end up creating useless nodes in the graph. The solution is ok for
 > now, but alternatively we could have a call GetAccountingEcForTask
 > which returns the ec used for accounting. This ec will not be
 > returned by GetTaskEC in the Whare-M cost models, but it could be
 > returned in other cost models. In this ways we avoid having
 > superfluous nodes.

Agreed, although I suspect a more general way of doing this would be to add a tag to each EC that denotes whether we want it to be represented in the flow graph. This will require us to move to representing ECs by more than an uint64_t ID, but I've been thinking that we'll need to do that anyway (for various other reasons).

Patch-set: 1
Label: Code-Review=0