Fix two bugs with ECs for first task out of a larger job.
When using the Whare-Map cost model and submitting jobs with >1 task,
two bugs related to the creation of ECs occurred prior to this change:

 1. Because we compute the has corresponding to the per-program TEC
    in two places and those notions had got out of sync, tasks would
    only have edges to ECs added via AddTaskEquivClass in the
    FlowGraph, rather than via GetIncomingEquivClassPrefArcs. This
    affected the first task (triggering EC creation) especially.
 2. TECs ignored by the GetIncomingEquivClassPrefArcs logic in the
    cost model (e.g., the new per-job EC) would not be connected
    to these tasks either, but would be connected to others (via
    AddTaskEquivClass).

The change fixes these bugs by (a) bringing the notions back in
sync by basing GetIncomingEquivClassPrefArcs on the set of ECs
as computed by GetTaskEquivClasses; and (b) adding an arc from
AddTaskEquivClass in FlowGraph even after EC creation (but only if
it does not already exist).

In addition, the change also brings in a number of safety checks
in various places dealing with arc addition, to guard against
duplicate arcs.

Change-Id: If773062a55c0200a8eb85c4e9328d201165ad7eb