Only inject TaskLib monitor thread into actual task binaries.
This change fixes a long-standing bug related to executing wrappers
around the task binary (e.g., gdb or perf): when executing within
a wrapper, both the wrapper binary and the task binary would start
monitor threads and all messages would be duplicated. This is
generally fine, as the messages are idempotent, but could occasionally
lead to tricky races and wedging. With this change, monitor threads
are only started for processes whose /proc/self/comm is identical to
their binary name (generally the case on Linux).

Affected modules: engine

Change-Id: I68b99b019d350bda3cb0b25eed31ad742ba69803