Correctly handle failed remote task delegation.
This change fixes a bug related to handling failed delegation of a
task to a remote executor. This can happen because a remote resource
has failed, or because another scheduler placed a task on the same
resource in the meantime (a "conflict" in Omega terms). Previously,
we didn't handle this case correctly, and the task was lost. Now,
we first put delegated tasks into the ASSIGNED state, and only move
them to DELEGATED once the far end has confirmed that the delegation
succeeded. If delegation fails, the task is moved back to RUNNABLE
state and the scheduler is invoked to try again placing it.

Note that we do not currently have any protection against infinite
failed delegation loops; this won't in practice, though.

Affected modules: engine scheduling

Change-Id: Ic3c102b558f5187180c1f0708d09f77b835dd239