Don't clean up flow graph state for delegated tasks.
This change fixes a bug that would occur when delegating a task
to a coordinator that runs the flow scheduler. Once the delegated
task completes, the coordinator would attempt to clean up
non-existent flow graph state for the task. Instead, we now skip
this clean-up step for delegated tasks.

In the future, we may want to re-visit the implications for local
scheduling decisions made by the subordinate flow scheduler.

Affected modules: scheduling/flow

Change-Id: Ibaf712a0991e4497d3ca4ca4795a62479a45ac69