Add COORDINATOR node type and use it in FlowGraph.
This change replaces the GLOBAL_AGGREGATOR flow node type (which we
no longer use) with a COORDINATOR one. This is used to represent the
resource corresponding to a coordinator in the flow graph; the CoCo
cost model requires this (as the COORDINATOR -> MACHINE arcs contain
machine-specific costs).

Affected modules: scheduling/flow

Change-Id: I893175be58351723f395f0aab27ab661a0550f81