Propagate interference scores across topology.
This change implements the CoCo cost model's propagation of
interference scores from the tasks (at the leaves) through the
resource topology. Compared to the written description, a number of
bugs have been fixed:
 1) The scale factor's directionalty is now correct (more idle cores
    = lower scale factor = lower cost).
 2) The impact of a particular leaf's score *decays* through the
    topology, as the formula calculates, for vertex u with children
    v \in V:

      cost(u) = ( e^( (t-i) / t) * \sum_v( cost(v) / |V| ) ) -
                \sum_v( cost(v) / |V| )

    This leaves only the cost execess caused by non-idle leaves, with
    an impact that is proportional to the fraction of occupied leaves,
    but which decays in magnitude (although costs are, of course, summed
    all the way to the leaves when solving the optimisation).

In order to make this work, we now track both the number of occupied
leaves below a resource and the total number of leaves below it via the
resource statistics mechanism (similar to how the Octopus cost model
works).

Affected modules: base engine scheduling/flow

Change-Id: I5af001d518176b1275b7e1c32936ed68e613161a