'Correctly' track available memory.
Previously, or notion of "free RAM" excluded the buffer cache. This
change includes the buffer cache memory in the available RAM. In order
to do this, we have to change the way we extract memory information from
the system: sysinfo(2), which we previously used, does not return an
indication of buffer cache size. Instead, we now parse /proc/meminfo,
which does have the information.

Affected modules: platforms/unix#

Change-Id: I9c85574b25d499c469418cc9bcb710522f79891f