Regularly re-visit arcs to and from ECs.
This change adds functionality to update the set of arcs from and to an
equivalence class node by re-visiting it after creation. Previously, the
set of arcs to/from an EC aggregator and their costs would be fixed at
creation time. Instead, we now re-visit each EC when AddOrUpdateJobNodes
is invoked (e.g. when updating time-dependent costs) and add additional
arcs if needed, or adjust the cost on existing ones according to the
cost model.

This solution (i) does not currently delete arcs that have been removed
from the EC's incoming/outgoing arc sets as indicated by the cost model;
and (ii) potentially makes many redundant visits to each EC node if many
tasks have arcs to it. We should address both of these points in the
future.

Affected modules: scheduling/flow

Change-Id: I450bbcf5d0e25cc335698acaa2aa1d9ad4736dec