Correctly track min and max machine capacities.
This change adds functionality to the CoCo cost model to dynamically
track the maximum and minimum machine capacity throughout the cluster
in each dimension. We need this because some of the costs in the CoCo
cost model depend on these values. Furthermore, this change adapts the
costs used in the CoCo cost model in the following ways:
 1) Costs within the resource topology (which represent load) are
    normalised to the machine capacity.
 2) The cost from a task node to its TEC (which expresses resource
    requirements) is normalised by the maximum machine capacity in the
    cluster (i.e., it expresses the best-case load added).
 3) The cost from a task to its unscheduled aggregator is normalised
    by the minimum machine capacity in the cluster (i.e., it expresses
    the load added when scheduling in the most loaded place).

Affected modules: scheduling/flow

Change-Id: Ib120876e7750fe4a422c4344ab921cb360f3d52a