Correctly propagate and assign interference scores.
This change completes the CoCo cost model implementation by propagating
the interference score vectors (now represented by the
CoCoInterferenceScores protobuf) through the topology and by assigning
the interference score on the TEC -> resource aggregate arc as the
vector's element in the incoming task type dimension. (In other words,
the cost is the penalty for running a task of this type under this
resource aggregate). When propagating the interference vector, we
divide by the number of subordinate resources at each branching point in
order to keep the cost within (0, omega_). However, the division only
takes place when using the accumulated value *at the next level*, since
we don't know whether we have accumulated to the total before then.

Affected modules: base scheduling/flow

Change-Id: I064bfb37d0e0944f5b5ff82a494b5f01d18c989f