Count how often a task fits in resource aggregate.
This change adds the necessary methods to count how many times a task's
resource request vector fits into the available resources under a
resource aggregate. In the future, we will use this information to set
the capacity on a TEC -> resource aggregate arc in the CoCo cost model.

Affected modules: scheduling/flow

Change-Id: I845a22f5bf7694553ee01fb1f74a1e5e438d07f9