Cost models can specify capacity on EC arcs.
This change adds the ability for cost models to specify both the cost
and the capacity on outgoing arcs from equivalence class aggregators to
resource nodes. This is an API change: instead of returning just a
Cost_t, EquivClassToResourceNodeCost() now returns a pair<Cost_t,
int64_t>, with the first element being the cost and the second element
being the capacity. If the capacity is set to a negative value, the flow
graph automatically uses the maximum possible capacity.

The change also updates all of the cost models to use the new API.

Affected modules: scheduling/flow

Change-Id: I2972bae4567d3fdfac1fd68f388014e94e452ae4