Allow plotting an individual flow graph.
Previously, the script to plot flow graphs using dot would always plot
*all* of the flow graphs in the debug directory. This change allows
passing an argument to specify which flow graph is to be plotted. This
makes it easier to debug flow graphs when there is a large number of
them. The change also adapts the plotting hack in the HTTP UI to make
use of this feature.

Affected modules: scripts engine

Change-Id: I15f3723a42767f3449296119f1f57bc01d626d6d