Close files after serving via web UI.
This change fixes a bug in the web UI back-end: when serving a file
via the HTTP UI, we did not close it afterwards. This ended up leaking
FDs (up to hundreds in the worst case). Alongside, I also fixed the
error condition and increased the read buffer size.

Affected modules: engine

Change-Id: I08f1a4e4d0e636de317a612dd6e32ca8d5a81d71