Add EC arcs to running tasks only with preemption.
This change fixes an issue that introduced unprocessed preemption
and migration deltas even if preemption was turned off. We previously
added arcs to EC aggregators for tasks in both RUNNABLE and RUNNING
states. However, arcs for RUNNING tasks are only required if preemption
is enabled. This change introduces an appropriate check to all cost
models that use EC aggregators.

Affected modules: scheduling/flow

Change-Id: I609375ba9c42bcdd79460b6da98af52d6c0688dc