Fix race condition in scheduler implementations.
This change fixes a nasty concurrenty bug that occured when a newly
submitted job raced a task completion for the scheduler lock. Previous
code, based on assumptions made when the scheduler was protected by a
boost::mutex rather than a boost::recursive_mutex, only took locks out
*after* calling the superclass handlers (e.g. FlowScheduler calling into
EventDrivenScheduler methods), since they would otherwise deadlock. The
superclass methods themselves obtained the scheduler lock on invocation.
However, this exposes a brief period in between the methods where the
lock is released, and can be obtained by another thread (e.g., on the of
the HTTP server threads handling a newly submitted job). The result was
a race that led to stale state being used in the FlowScheduler, and a
corresponding failure once the bogus flow graph had been solved and the
deltas were to be applied (typically, scheduling a no-longer-existent
task).

The recursive mutex fixes the problem, since it can be taken and held by
a thread for the entire duration of processing in the scheduler, until
control is returned to the coordinator.

Affected modules: scheduling

Change-Id: If0f1bc5aa0e9525c388036b3f95a9ac3f4176d85