Check if sockets are open before closing them.
This change fixes a nasty concurrency bug owed to the fact that UNIX FD
numbers are re-used after closure. A race on closing the socket used to
communicate with a task could (very occasionally) lead to the FD being
closed twice (once when the TCP connection drops and once when we tear
down the channel). If the FD number had been re-used for something else
in the meantime (a few hundred microseconds), we would inadvertently
close that FD. As a result, web UI HTTP connections sometimes dropped,
and the solver pipe sometimes failed.

This change fixes the bug by properly checking if sockets are still open
before closing them (we previously assumed that ASIO does this; it does
not).

Affected modules: platforms/unix

Change-Id: I0a36c738cb6ad6688b030094b9c32dbe9176cdbf