Make more of the scheduler methods virtual.
We support multiple pluggable schedulers, each implementing the
interface in SchedulerInterface. Typically, schedulers sub-class the
EventDrivenScheduler class. However, many of the API methods were not
declared as virtual, making it unclear whether the subclass method
(which typically upcalls into EventDriveScheduler) is invoked, or
whether the EventDrivenScheduler method is invoked directly. (In
practice, this wasn't a problem: we almost always use explicit types for
the scheduler.)

Affected modules: scheduling

Change-Id: Iada212273b17313ff60ed9a5ce4802956f7fd234