Delete stale outgoing arcs from EC to resources.
This change addresses an issue exhibited by resource-fitting in the CoCo
cost model: arcs to suitable nodes might be added when resources are
available, but for various reasons a task may not immediately schedule
using this arc. If another task invalidates the premises of the arc
addition (by e.g., using available resources), the arc still remains,
enabling the task to later scheduler, but violate the resource fit
criteria. This change fixes the problem by computing the difference
between the existing outgoing arc set and the computed arc set at the
current point in time, removing any arcs in that difference set.

Affected modules: scheduling/flow

Change-Id: Icfce1b670873a085a4fa0c007c0217745b5c0066