Add initial support for resource reservations.
This change introduces the infrastructure necessary to support resource
reservations. However, they won't work quite yet: code to set and propagate
the reservation will be pushed in a separate change.

The current implementation of reservations is *always strict*: in other
words, a task will never be placed in a location where the resource
reservation is not satisfied in all dimensions. In principle, however,
the code is set up to support non-strict fit (opportunistic overcommit)
and partial fit.

Affected modules: base scheduling/flow

Change-Id: If503ae17daa51129d45a61ceb0c84b958d500237