Use an MRSW mutex to protect the KB.
This change fixes a concurrency bug in the knowledge base code.
Previously, the knowledge base returned pointers to collections (e.g.,
the deque for machine samples) that could be mutated "under the caller's
feet", causing occasional segfaults. The protection by kb_lock_ was
inadequate since it only protected write accesses.
This change instead uses an MRSW mutex that protects both reads and
writes, and changes the API to return copies of data instead of pointers
(necessary since the lock cannot be held post-return). The change also
adds a convenience call for the common "get latest sample" case, copying
less data in this instance, and updates all consumers of the API to use
it correctly.

Affected modules: engine scheduling

Change-Id: I2a735f037eb7ee17ea9e60470de758182e5636f8