Fixed logic to compute the fraction of events to retain.
This change fixes several bugs in the trace simulator. First, it
fixes the way of computing the event discard threshould. Previously,
the computation could overflow uint64_t due to the way doubles are
represented. Secondly, it fixes a conditional used for validating
the flags. Lastly, it extends the destructor to free up some memory
the code was previosly leaking.

Affected modules: sim/trace-extract

Change-Id: I8e27da4c2935d4cfb580e22e8a72092496233cb7