Improves the code that randomly picks a value from an unordered_set.
This diff changes the code that selects a random element from an
unordered_set. The code was previously picking an index within the
set and then moving an iterator to that index. The new code first
picks a non-empty bucket of the set and then randomly selects an
element within that bucket. With this change the complexity of the
code changes from O(number of elements in set) to
O(number of elements in bucket).

Affected modules: misc scheduling/flow

Change-Id: I997baf504f456af2ca3a855a871a2363eeb78a87