Update review board URL and run instructions.
This change updates the README file to point to the board of open code
reviews, rather than those which are already merged. In addition, it
adds a sentence about having to run the coordinator from the Firmament
root directory in order for the web UI to find its template files.

Change-Id: Ice37d2be0c346983c22c4a32eb04bbc1ea23a475