Fix build on Ubuntu 12.04.
This change fixes the build issues experienced on Ubuntu 12.04 after the
inclusion of libhdfs3. To solve the dependency on C++11 std::chrono,
which isn't available in the gcc C++ stdlib on 12.04, we now build
libhdfs3 with Boost support (as we have Boost 1.55 packages available on
Ubuntu 12.04). The change also changes the build directory for libhdfs3
to be in line with the other external library dependencies.

Change-Id: I62d56028cd6a4d0cb53bbce7d8e4812b1a0e41cf