Re-factor simulator so that it can run non-Google simulations.
This change improves the simulator so that the code that loads
the task events from the trace is not intertwined with the
code that plays the simulation. After this change we will be
able to generate synthetic simulations.

This change also replaces every occurence of solver with
scheduler.

Affected modules: scheduling sim

Change-Id: Ia9c5c820ae24f6ce61903f636fdc064fea4f8391