Added TraceLoader interface to decouple trace processing from loading.
This change adds a TraceLoader interface that must be
implemented by any simulation. The TraceLoader is responsible
for adding the events the simulator plays.

Whenever we will want to add a new type of simulation we will just
have to define a new TraceLoader (see SyntheticTraceLoader for
an example).

Affected modules: sim

Change-Id: I82fee23b32a32ea0931422c2168cfee95b5b7590