Fixed simulator's event filtering.
Previously, the simulator was filtering events by applying a hash
over the TraceTaskIdentifier structure. However, this could
lead to scenarious in which jobs would not complete because
the simulator would wait for filtered tasks to complete.

This change fixes the issue by only passing the job id to the
hash function.

Affected modules: sim

Change-Id: I3c7d793d73a6538772cf8eb0345d3eefa43a21d0