Fixed simulator's event filtering.
Previously, the simulator was filtering events by applying
a hash over the TraceTaskIdentifier struct. However, this could
lead to scenarios in which jobs would not complete because the
simulator would wait for filtered tasks to complete.

This change fixes this issue by updating the number of tasks
a job has whenever a task event is filtered.

Affected modules: sim

Change-Id: I5c1f5eea5c394a4b11e8870f091ae96c5e944223