Fix bug in WhareMap and Octopus cost models.
This change fixes a bug that occurs in the initialization phase of the
scheduler when running with the WhareMap and Octopus cost models. Since
the resource topology isn't fully initialized when GatherStats() runs
for the first time, the ResourceStatus and ResourceDescriptor pointer
can be NULL at this point. Instead of crashing, we now ignore the error
(as we do in other cost models). This is safe because GatherStats() will
simply return early, but is guaranteed to run again when any
scheduling decisions are made.

Issue: #35
Affected modules: scheduler/flow

Change-Id: Idab5e28ff49874337f4ded4f2ebf62cec1460f76