Updated arc changes to include old cost and arc type.
This change updates the extended DIMACS arc change generated
by Firmament to include arc's type (i.e. OTHER or RUNNING arc)
and its old cost. The old cost is used to detect the exact arc
that is going to be changed among the possibly multiple arcs
between the same pair of nodes.

Change-Id: I74b9f7d3d2107ab9370f937043bcb1a68c988ee2