Update GenerateTrace to print trace ids when running in simulation mode.
This feature is implemented in the following way:
1) For every job the simulator passes its trace job id by setting the
id as the job name.
2) The GenerateTrace class checks whenever it prints a job trace entry
if the job has set in its name a trace id. If it has, then it prints
the trace id instead of the Firmament id.

The change implements the same behaviour for trace machines as well.

Affected modules: misc scheduling/flow sim

Change-Id: I3317a620c89002aa7769f34c901788df1dc5e166