Updated Firmament to be able to run with real or simulated time.
This change updates the code so that schedulers and cost models
can work with simulated time. Some cost models make use of time when
they adjust the cost of the arc from a task to an unscheduled aggregator.
Thus, when running in simulation mode this adjustment would not be accurate
unless the cost models use simulation time to compute the cost update.

Moreover, GenerateTrace prints a timestamp of any event it prints. Previously,
it was using the real time which was not accurate.

This change adds a new TimeInterface that offers a method to get the current
timestamp. Depending on the mode in which we run we pass a simulated Timer
or a RealTimer into the constructors that need access to a timer. In this way
we can create an accurate simulation that works with the correct timestamps.

Affected modules: engine engine/executors misc scheduling scheduling/flow sim

Change-Id: I99b9063a9ec8b39e7f1b41d28c4142473ff0051f