Update simulation time before and after solver run.
This change does the following changes:
1) It makes sure that the simulation timestamp is updated before a solver run to
the timestamp of the beginning of the solver run. Hence, the events that
happen as a result of a scheduling now have correct timestamps.
2) Updates the simulation timestamp to the value before the solver run to make
sure that trace events that happened in the meantime are correctly replayed.
3) Changes SchedulerStats to use u-sec rather than seconds. The entire simulator
uses u-sec, hence it makes sense to be consistent in SchedulerStats as well.
4) Updates the code to call JobCompleted when all the tasks of a job complete.

Affected modules: misc scheduling/flow scheduling/simple sim

Change-Id: I7e8ec6203edd757b655f6b0104ad5602ba446bb6