Generate trace with the correct task ids when running in simulation mode.
This change updates the code to pass the trace task ids to
generate_trace such that the output trace contains the same task ids
as the input trace.
The implementation is similar to the one for passing trace job ids.
We set the trace task id as the index of the task. We can't directly
set it as the id of the task because the trace task ids are only unique
per job.
In generate_trace we detect when a submitted task is a simulation job and
in such case we fetch the task id from the task descriptor's index. This
is then updated in the TaskRuntime structure which is used by the other
methods to log the events with the correct task id.

Affected modules: misc scheduling/flow sim

Change-Id: I7be1d74c9b6bc19a4d3d3c36509613fbfc499830