Updated the scheduler to play simulation events that happen while the solver is running.
This change does the following:
1) Adds a OnSchedulingDecision callback to the scheduler event notifier.
This callback is used by the simulator to replay the events that happened
from the start of the solver run until its end. The simulator playes the
events befure the scheduler applies the decisions it has made.
2) Tasks may complete or PUs may be removed during the solver run. We now
handle this case using two sets in which we store the tasks that
have completed and the PUs that have been removed while the solver was
running. After the end of the solver run, we make sure not to place
completed tasks or tasks on PUs that have been removed.
3) RunnableTasksForJob was previously returning a set containing the
runnable tasks for every job. The change now fixes this by using a
global runnable_tasks_ that stores runnable tasks on a per job basis.
4) Removes obsolete scheduler statistics printing.

Affected modules: scheduling scheduling/flow scheduling/simple sim

Change-Id: I764fa8dd02ab02b35cf4331ddf1013e82de31088