Updated solver_dispatcher to read the algorithm runtime from DIMACS changes.
This change updates solver_dispatcher to read the duration of the algorithmic
part of the solver from the DIMACS format. Previously, we were parsing
stderr of the solver for "ALGO TIME". However, this solution was requiring a
flag to be manually set for the solvers that were printing the algorith time.
With this new change we don't need the flag. We just check if there is a
comment "C ALGORITHM TIME duration" in the DIMACS exchange between the
solver and Firmament.

Change-Id: I2572cd2d918e282b440b8af02fc2b1c1f4469adf