Combined prepare, gather and update resource stats in a single traversal.
This change improves the performance of UpdateCostModelResourceStats by
merging all the passes into a single one. Moreover, the change redudes the
complexity of the function from MlogN to M by storing a visited field
in FlowGraphNode rather than using a set to mark the processed nodes.

Affected modules: scheduling/flow

Change-Id: I85b194d7f7fac82e252f359f15cdf324c548d4c0