Improved the performance of NodeForResourceID.
Previously, NodeForResource would required two lookups and it would
account for about 26% of the time of Firmament's side of simulation.
This change replaces FlowGraphManager's resource_to_node_id map with
a map from resource ids to FlowGraphNodes. In this way we half the
number of lookups.

Affected modules: scheduling/flow

Change-Id: Iaf3879e0fe8b0f4c1cafb9ab4b2fe55db0a083e4