Pass trace job id using TaskDescriptor's name field.
This change uses the TaskDescriptor's name field to pass the
job trace id from the simulator to the TraceGenerator. Previously,
we were using the name field of the JobDescriptor. However, this
solution was requiring us to pass JobDescriptors to the TraceGenerator.
As a result, in AddOrUpdateJobNodes we could only consider one job at
a time which would lead to big inefficiencies.

Affected modules: misc scheduling/flow sim

Change-Id: I08ce9973a90fd2f54b48c520f98ed7e69a89dfad