Improved AddOrUpdateJobNodes by removing calls to GetIncomingEquivClassPrefArcs.
This change removes the inefficient GetIncomingEquivClassPrefArcs
method. Previously, for each task we would get the equivalence
classes and then for each equivalence class we would get the
incoming arcs by using GetIncomingEquivClassPrefArcs.
Instead, we now just get the equivalence classes of a task
and update the arcs from the task to the equivalence classes.

Affected modules: scheduling/flow

Change-Id: I0af11d63570e6f366b8669dca57f67884e96caa3