Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

LGTM, although see below.

This change, if I'm not mistaken, does not result in a reduction in functionality in practice, but does amount of a reduction in theory. Specifically, it was previously possible to write a cost model that does NOT have arcs from each task to ALL of its equivalence class aggregators. For example, a cost model could dynamically decide not to add arcs to a specific EC aggregator due to some global condition. We don't require that sort of thing in any of our existing cost models, but the new code will always add arcs from each task to all of its EC aggregators. That's probably fine, but we should think a little about whether there is ever a use-case for the above notion of selectively adding arcs (which would require a call similar to GetIncomingEquivClassPrefArcs to exist).

Patch-set: 1
Reviewer: Malte Schwarzkopf <1004901@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2
Label: Verified=+1